Update to new preprocessor style
