+++ docs/book.toml
@@ -5,8 +5,14 @@ multilingual = false
 src = "src"
 title = "Frequency Developer Gateway"
 
-[preprocessor.local]
-command = "node preprocessor.mjs"
+[preprocessor.button-links]
+command = "node preprocessors/button-links.mjs"
+
+[preprocessor.svg-embed]
+command = "node preprocessors/svg-embed.mjs"
+
+[preprocessor.swagger-embed]
+command = "node preprocessors/swagger-embed.mjs"
 
 
 [output.html]
