+++ docs/preprocessors/README.md
@@ -1,0 +1,20 @@
+# Preprocessors
+
+Folder for various preprocessors for Gateway Documentation.
+
+## Setup
+
+1. Start with one of the existing mjs files.
+2. Duplicate with a unique kebab case name.
+3. Add it to book.toml.
+
+## Resources
+
+- List of many of the already existing mdbook plugins: https://github.com/rust-lang/mdBook/wiki/Third-party-plugins
+- https://rust-lang.github.io/mdBook/format/configuration/preprocessors.html
+- https://rust-lang.github.io/mdBook/for_developers/preprocessors.html
+
+## Troubleshooting Help
+
+- In JavaScript, use console.error, not console.log
+- Make sure your exit codes
